[fix] Ignore 5xx responses from Kuvera
If we get a 5xx on a batch, split that batch into two and continue
till we find a failing batch with count=1 - mark it as failed.

We run this on batches of 150, so we effeciently try to make 30 calls.
This increases the overhead somewhat, by adding ~8 calls to find the
broken identifier which returns a 5xx. But still, we get the whole
lookup table within 40 API calls, so that's quite good.